requirements.txt
bcrypt
rapidfuzz
sqlalchemy
flask
pymysql
Werkzeug
python-dotenv
